+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.15.2-1.X.Y.Z] - Not released
 ### Added
 - Using glowstone on the Dye Liquifier to swap internal tank mode between dump/static
+- Added some basic TOP support
+- Updated JEI plugin to display fluids a bit more clearly and show Pure Dye recipe
 ### Changed
 - Fix internal displaying missing block as (x,z,z) instead of (x,y,z) [#436]
 - Remove all entities in the simulation cell after sweeping the drops [#419]